Update bookHandler.java , 2 issue solved.

+++ BookStoreApi/src/main/java/org/example/Controller/bookHandler.java
@@ -47,12 +47,18 @@ public class bookHandler implements HttpHandler {
             }
             else{
                 var headers = exchange.getRequestHeaders() ;
-                Long id = parseLong(headers.get("id").get(0));
-                String  name = headers.get("name").get(0) ,
-                        author = headers.get("author").get(0) ,
-                        genre = headers.get("genre").get(0) ;
 
-                response.append(bookService.saveBook(new Book(id , name , author , genre)))   ;
+                if( ! (headers.containsKey("id") && headers.containsKey("name") && headers.containsKey("author") && headers.containsKey("genre") ) ){
+                    response.append("Headers Missing") ;
+                }
+                else{
+                    Long id = parseLong(headers.get("id").get(0));
+                    String  name = headers.get("name").get(0) ,
+                            author = headers.get("author").get(0) ,
+                            genre = headers.get("genre").get(0) ;
+
+                    response.append(bookService.saveBook(new Book(id , name , author , genre)))   ;
+                }
             }
         }
         else if(method.equals("DELETE")){
@@ -78,15 +84,20 @@ public class bookHandler implements HttpHandler {
             else{
                 String query = uri.getRawQuery(); // uri er question er por ja ce
                 Long oldId = parseLong(query) ;
-
                 var headers = exchange.getRequestHeaders() ;
-                Long id = parseLong(headers.get("id").get(0));
-                String  name = headers.get("name").get(0) ,
-                        author = headers.get("author").get(0) ,
-                        genre = headers.get("genre").get(0) ;
 
-                bookService.deleteBook(oldId);
-                response.append(bookService.saveBook(new Book( id, name , author , genre)))   ;
+                if( ! (headers.containsKey("id") && headers.containsKey("name") && headers.containsKey("author") && headers.containsKey("genre") ) ){
+                    response.append("Headers Missing") ;
+                }
+                else{
+                    Long id = parseLong(headers.get("id").get(0));
+                    String  name = headers.get("name").get(0) ,
+                            author = headers.get("author").get(0) ,
+                            genre = headers.get("genre").get(0) ;
+
+                    bookService.deleteBook(oldId);
+                    response.append(bookService.saveBook(new Book( id, name , author , genre)))   ;
+                }
             }
         }
